+++ packages/gamut-illustrations/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.54.3](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-illustrations@0.54.2...@codecademy/gamut-illustrations@0.54.3) (2025-04-25)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
+### [0.54.2](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-illustrations@0.54.1...@codecademy/gamut-illustrations@0.54.2) (2025-04-25)
+
+**Note:** Version bump only for package @codecademy/gamut-illustrations
+
 ### [0.54.1](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-illustrations@0.54.0...@codecademy/gamut-illustrations@0.54.1) (2025-03-27)
 
 **Note:** Version bump only for package @codecademy/gamut-illustrations